+++ geo/CHANGES.md
@@ -38,6 +38,11 @@
   Haversine::distance(p1, p2)
   ```
   * <https://github.com/georust/geo/pull/1216>
+* Change IntersectionMatrix::is_equal_topo to now consider empty geometries as equal.
+  * <https://github.com/georust/geo/pull/1223>
+* Fix `(LINESTRING EMPTY).contains(LINESTRING EMPTY)` and `(MULTIPOLYGON EMPTY).contains(MULTIPOINT EMPTY)` which previously
+  reported true
+  * <https://github.com/georust/geo/pull/1227>
 
 ## 0.28.0